pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pwn-solver"
version = "0.1.0"
description = "Agentic LLM system for binary exploitation, powered by MCP tool servers"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "anthropic>=0.42.0",
    "mcp>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
pwn = [
    "pwntools>=4.12.0",
    "ropper>=1.13.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[project.scripts]
pwn-solver = "agent.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
# Ghidra post-scripts are Jython inside Ghidra's VM (injected globals); not plain Python for Ruff.
exclude = ["mcp-servers/exploit-tools/ghidra_scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"scripts/fsop_shell_solve.py" = ["F403", "F405"]
"agent/prompts.py" = ["E501"]
"agent/tools.py" = ["E501"]
"mcp-servers/__init__.py" = ["N999"]
"mcp-servers/dynamic-analysis/__init__.py" = ["N999"]
"mcp-servers/exploit-tools/__init__.py" = ["N999"]
